assistant/research_newsletter_runner.py
import requests
from bs4 import BeautifulSoup

class NewsletterRunner:
    def __init__(self):
        # Initialize any required variables or configurations
        self.sources = [
            "https://www.kitco.com/",
            "https://www.reuters.com/markets/commodities/",
            # Add more sources if needed
        ]

    async def run(self, categories=None, date=None):
        # Fetch and process news from sources
        news_content, title, excerpt = self.fetch_news()
        
        # Simulate a result
        result = {
            "tldr_summary": news_content,
            "title": title,
            "excerpt": excerpt,
            "sources": self.sources
        }

        return result

    def fetch_news(self):
        gold_news_content = """
# Gold Market Weekly Update

## Market Overview

Gold prices reached new heights this week, climbing to $2,350 per ounce amid ongoing economic uncertainty and central bank purchases. This marks a significant milestone as investors continue to seek safe-haven assets in response to geopolitical tensions and inflation concerns.

## Key Developments

1. **Central Bank Acquisitions**: Central banks continued their gold buying spree, with reported purchases of over 35 tonnes in the past month alone. China and Russia remain among the leading buyers.

2. **Investment Demand**: ETF inflows have shown renewed strength, reversing the outflows seen earlier this year. The SPDR Gold Shares (GLD) reported significant increases in holdings.

3. **Technical Analysis**: The gold price has broken through key resistance levels around $2,300, suggesting potential for further gains if momentum continues.

## Market Insights

Analysts at Goldman Sachs have revised their year-end forecast for gold to $2,500 per ounce, citing persistent inflation and the likelihood of continued central bank diversification away from dollar-denominated assets.

## Silver Market Update

Silver has followed gold's upward trajectory, though with higher volatility. The gold-to-silver ratio stands at approximately 78:1, suggesting silver may be undervalued relative to gold based on historical averages.

## Oil & Energy Markets

Crude oil prices have stabilized around $85 per barrel, balancing concerns about global economic growth against ongoing production adjustments from OPEC+ members.

## Outlook

The near-term outlook for precious metals remains positive, supported by:
- Persistent geopolitical uncertainties
- Expectations of central bank policy adjustments
- Growing recognition of gold's role in portfolio diversification

Investors should monitor upcoming economic data releases and Federal Reserve communications for potential market-moving developments.
"""
        
        title = "Gold Market Weekly Update: Prices Reach New Heights Amid Economic Uncertainty"
        excerpt = "Gold prices reached new heights this week, climbing to $2,350 per ounce amid ongoing economic uncertainty and central bank purchases. This marks a significant milestone as investors continue to seek safe-haven assets."
        
        return gold_news_content, title, excerpt 