requirements.txt
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachelib==0.13.0
cattrs==24.1.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dominate==2.9.1
Flask==3.1.0
flask-babel==4.0.0
Flask-Bootstrap==3.3.7.1
Flask-Caching==2.3.1
Flask-WTF==1.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
platformdirs==4.3.6
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
requests-cache==1.2.1
six==1.17.0
soupsieve==2.6
typing_extensions==4.12.2
url-normalize==1.4.3
urllib3==2.3.0
visitor==0.1.3
Werkzeug==3.1.3
WTForms==3.2.1
